Refactor BlogPost into separate component with file upload and publish functionality
